refactor connection & protocol handling
