+++ libmproxy/proxy/connection.py
@@ -68,7 +68,15 @@ class ClientConnection(tcp.BaseHandler, stateobject.StateObject):
         return f
 
     def convert_to_ssl(self, *args, **kwargs):
-        tcp.BaseHandler.convert_to_ssl(self, *args, **kwargs)
+        def alpn_select_callback(conn_, options):
+            if alpn_select in options:
+                return bytes(alpn_select)
+            else:  # pragma no cover
+                return options[0]
+
+        # TODO: read ALPN from server and select same proto for client conn
+
+        tcp.BaseHandler.convert_to_ssl(self, alpn_select=alpn_select_callback, *args, **kwargs)
         self.timestamp_ssl_setup = utils.timestamp()
 
     def finish(self):
@@ -160,7 +168,10 @@ class ServerConnection(tcp.TCPClient, stateobject.StateObject):
                 self.address.host.encode("idna")) + ".pem"
             if os.path.exists(path):
                 clientcert = path
-        self.convert_to_ssl(cert=clientcert, sni=sni, **kwargs)
+
+        # TODO: read ALPN from client and use same list for server conn
+
+        self.convert_to_ssl(cert=clientcert, sni=sni, alpn_protos=['h2'], **kwargs)
         self.sni = sni
         self.timestamp_ssl_setup = utils.timestamp()
